requirements.txt
Flask
SpeechRecognition
gTTS
pygame
python-dotenv
markdown
requests
